Package move

+++ src/main/java/com/axa/sesummit/codinggame/services/RomanPriceService.java
@@ -2,5 +2,5 @@ package com.axa.sesummit.codinggame.services;
 
 public interface RomanPriceService {
 
-	public Double romanPrice(Integer days);
+	public Double romanPrice(Long days);
 }
